scripts: runners: add 'bsim_device' and 'bsim_phy' runners
Those are runners that handle 'west flash' request in context of sysbuild.
Instead of executing binary directly, all command arguments are collected
to be executed with the last sysbuild domain. That way all BabbleSim
components can be executed at once and hence communicate with each other
within a single session.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>