scripts: runners: add 'bsim_zephyr' runner
This is a slight modification of 'native' runner, with capability of
automatically running BabbleSim with 2G4 Phy and optionally other
components for Bluetooth simulation.

Override default FLASH runner, so that 'bsim_zephyr' is used instead of
'native' runner. That way execution in context of sysbuild with be
postponed till the last sysbuild domain and then in parallel with all other
BabbleSim components include 2G4 Phy.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>