gateway: disable MBEDTLS_USE_PSA_CRYPTO by default
This makes sure that legacy mbedTLS crypto APIs are used for now, instead
of being enabled by default. So far only native_sim when compiled with BT
Host stack was selecting this option. As a result it was not possible to
connect to cloud due to missing crypto algorithms, which need to be
selected specifically for PSA crypto APIs (thsoe are distinct Kconfig
option).

In future this might be reverted, when SDK will automatically select proper
Kconfig options related to PSA crypto API.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>