Include empty app image in nRF5340 build on Thingy:91 X
As pointed out by Dan in #8, the controller build for the Thingy:91 X was
missing an app core image, which is required to be able to boot up the net
core properly. This likely didn't get detected in testing due to lingering
empty app images from previous builds. By adding a sysbuild config for the
controller, we can force an app image to be included in the sysbuild and
also get rid of the command line sysbuild config outlined in the readme.

Signed-off-by: Trond Snekvik <trond@golioth.io>