downlink: pull next block from queue immediately
When we finished reading a downlink block, we would free the block
and mark the current block in the downlink context as NULL. If
there was any space remaining in the buffer passed in by the caller,
then we would pop the next block off the queue and begin reading it.
However, if the number of bytes remaining in the block was the same
as the size of the buffer passed in by the caller, then the
current_block in the downlink context would remain NULL until the
next time the caller requested data (which would be after a round
trip GATT write to the peripheral node). During this time, if we
received another downlink block from the cloud, we would see that
the current block in the downlink context was NULL and assume that
the application was waiting for new blocks, so we would notify it
that new data is available. This caused an extra GATT write to get
queued, and because each GATT write response triggers a new GATT
write, any extra queued writes effectively permanently reduce the
capacity of the ATT buffers. When this happened enough times to
exhaust the ATT buffers, the downlink would fail.

Now, instead of marking the current block as NULL, we immediately
pop the next block off of the queue. This ensures that there is
always a valid block in the downlink context as long as there is
data available to be sent to the peripheral node.

Signed-off-by: Sam Friedman <sam@golioth.io>