downlink: notify on new blocks only after previous failed data request
Previously, we would notify clients of new blocks after any call to
`downlink_data_get()` that failed to fully fill the passed in buffer.
Now, we notify clients via the `data_available()` callback that new
blocks have been downloaded only after a call to `downlink_get_data()`
that fails to get any data at all.

This handles the situation where the client exhausts the block cache
but a new block is downloaded and added to the cache in the time that
it takes the client to transfer the data from the previous call. In
this case, we would both execute the `data_available()` callback and
the client would request more data as part of its normal operation
(i.e. after receiving the GATT write response). Now, if a block
is received while the client is transferring the data from the call
that exhausted the block cache, the transfer simply proceeds in the
normal course and the `data_available()` callback is not executed.

Signed-off-by: Sam Friedman <sam@golioth.io>