scripts: runners: bsim: support --foreground-domain
This option allows to select domain which will be executed in the
foreground. So far this was fixed to be the last one based on "flash"
order. From now on, default foreground process (when no --foreground-domain
is specified) is the default domain in sysbuild context.

This option doesn't change much as 'flash' runner, but it will be reused
for 'debug' runner, so that selecting domain under GDB will be done with
the same option.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>